pyproject.toml
[project]
name = "auto-fetch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "jinja2>=3.1.6",
    "openai>=2.16.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
]
